Consolidate Checksum classes into own package
