+++ app/src/debug/java/net/squanchy/support/checksum/CurrentTimeModule.kt
@@ -1,4 +1,4 @@
-package net.squanchy.support.injection
+package net.squanchy.support.checksum
 
 import dagger.Module
 import dagger.Provides
