+++ app/src/main/java/net/squanchy/service/repository/firestore/FirestoreTracksRepository.kt
@@ -5,7 +5,7 @@ import net.squanchy.schedule.domain.view.Track
 import net.squanchy.service.firebase.FirestoreDbService
 import net.squanchy.service.firebase.toTrack
 import net.squanchy.service.repository.TracksRepository
-import net.squanchy.support.lang.Checksum
+import net.squanchy.support.checksum.Checksum
 
 class FirestoreTracksRepository(private val firestoreDbService: FirestoreDbService, private val checksum: Checksum) : TracksRepository {
