+++ app/src/main/java/net/squanchy/support/checksum/ChecksumModule.kt
@@ -1,8 +1,7 @@
-package net.squanchy.support.injection
+package net.squanchy.support.checksum
 
 import dagger.Module
 import dagger.Provides
-import net.squanchy.support.lang.Checksum
 
 @Module
 class ChecksumModule {
